+++ src/screens/WalletScreen/components/ActionsPanel/ActionsPanel.utils.ts
@@ -2,10 +2,10 @@ const arr = [
 	{ name: 'Send', icon: 'sendStroke' },
 	{ name: 'Exchange', icon: 'exchangeStroke' },
 	{ name: 'Receive', icon: 'receiveStroke' },
-	{ name: 'Copy address', icon: 'copyStroke' }
+	{ name: 'Copy address', icon: 'copyStroke' },
 	// { name: 'New wallet', icon: 'walletStroke' },
 	// { name: 'Switch accounts', icon: 'avatarStroke' },
-	// { name: 'Delete wallet', icon: 'closeStroke' }
+	{ name: 'Delete wallet', icon: 'closeStroke' }
 ];
 
 export { arr };
