+++ src/screens/WelcomeScreen/Background/Background.styles.ts
@@ -1,0 +1,17 @@
+import { StyleSheet } from 'react-native';
+
+export default StyleSheet.create({
+	container: {
+		flex: 1
+	},
+	top: {
+		position: 'absolute',
+		right: 0
+	},
+	bottom: {
+		position: 'absolute',
+		bottom: 0,
+		right: 0,
+		zIndex: 0
+	}
+});
