+++ src/screens/WelcomeScreen/WelcomeScreen.styles.ts
@@ -1,22 +1,6 @@
 import { StyleSheet } from 'react-native';
 
 export default StyleSheet.create({
-	// ... BACKGROUND ...
-	backgroundContainer: {
-		flex: 1
-	},
-	backgroundTop: {
-		position: 'absolute',
-		right: 0
-	},
-	backgroundBottom: {
-		position: 'absolute',
-		bottom: 0,
-		right: 0,
-		zIndex: 0
-	},
-
-	// ... WELCOME SCREEN ...
 	container: {
 		width: '100%',
 		flex: 1,
